@@ -7,25 +7,29 @@
     "@jacobmischka/gatsby-plugin-copy-files": "^1.0.0",
     "date-fns": "^1.28.5",
     "flickity": "^2.0.9",
-    "gatsby": "^1.3.2",
-    "gatsby-link": "^1.0.1",
-    "gatsby-plugin-google-analytics": "^1.0.1",
-    "gatsby-plugin-offline": "^1.0.1",
-    "gatsby-plugin-react-helmet": "^1.0.1",
-    "gatsby-plugin-sharp": "^1.3.0",
-    "gatsby-plugin-styled-jsx": "^1.0.1",
-    "gatsby-remark-copy-linked-files": "^1.3.0",
-    "gatsby-remark-images": "^1.3.2",
-    "gatsby-remark-prismjs": "^1.2.0",
-    "gatsby-remark-responsive-iframe": "^1.3.0",
-    "gatsby-remark-smartypants": "^1.0.1",
-    "gatsby-source-filesystem": "^1.0.1",
-    "gatsby-transformer-remark": "^1.1.0",
-    "gatsby-transformer-sharp": "^1.0.1",
+    "gatsby": "^2.0.79",
+    "gatsby-link": "^2.0.7",
+    "gatsby-plugin-google-analytics": "^2.0.8",
+    "gatsby-plugin-offline": "^2.0.20",
+    "gatsby-plugin-react-helmet": "^3.0.5",
+    "gatsby-plugin-sharp": "^2.0.17",
+    "gatsby-plugin-styled-jsx": "^3.0.3",
+    "gatsby-remark-copy-linked-files": "^2.0.8",
+    "gatsby-remark-images": "^3.0.1",
+    "gatsby-remark-prismjs": "^3.2.0",
+    "gatsby-remark-responsive-iframe": "^2.0.8",
+    "gatsby-remark-smartypants": "^2.0.7",
+    "gatsby-source-filesystem": "^2.0.12",
+    "gatsby-transformer-remark": "^2.1.18",
+    "gatsby-transformer-sharp": "^2.1.10",
+    "prismjs": "^1.15.0",
+    "react": "^16.7.0-alpha.2",
+    "react-dom": "^16.7.0-alpha.2",
     "react-helmet": "^5.1.3",
-    "typeface-lato": "^0.0.35",
-    "typeface-open-sans": "^0.0.35",
-    "typeface-source-serif-pro": "^0.0.35"
+    "styled-jsx": "^3.1.3",
+    "typeface-lato": "^0.0.54",
+    "typeface-open-sans": "^0.0.54",
+    "typeface-source-serif-pro": "^0.0.54"
   },
   "license": "MIT",
   "main": "n/a",
@@ -35,10 +39,10 @@
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "devDependencies": {
-    "babel-eslint": "^7.2.3",
-    "eslint": "^4.3.0",
+    "babel-eslint": "^10.0.1",
+    "eslint": "^5.11.1",
     "eslint-plugin-react": "^7.1.0",
-    "stylelint": "^8.0.0",
-    "stylelint-config-standard": "^17.0.0"
+    "stylelint": "^9.9.0",
+    "stylelint-config-standard": "^18.2.0"
   }
 }
