@@ -1,26 +1,18 @@
-/* global graphql */
-import React, { Component } from 'react';
-import PropTypes from 'prop-types';
-import Link from 'gatsby-link';
-import Flickity from '../flickity.js';
+import React, { useEffect } from 'react';
+import { Link, graphql } from 'gatsby';
 
+import Layout from '../components/Layout.js';
+
+import Flickity from '../flickity.js';
 import { formatDate } from '../utils.js';
 
-export default class Post extends Component {
-	static propTypes = {
-		data: PropTypes.object
-	};
+export default function Post({ data: { markdownRemark: post } }) {
+	useEffect(() => {
+		createCarousels();
+	}, []);
 
-	constructor() {
-		super();
-
-		this.createCarousels = this.createCarousels.bind(this);
-	}
-
-	render() {
-		const post = this.props.data.markdownRemark;
-
-		return (
+	return (
+		<Layout>
 			<section className="post">
 				<header className="post-header">
 					<Link to="/blog" className="blog-link">
@@ -29,37 +21,30 @@ export default class Post extends Component {
 					<h1 className="post-title">
 						{post.frontmatter.title}
 					</h1>
-			{
-				post.frontmatter.date && (
-					<time className="post-meta">
-						{formatDate(post.frontmatter.date)}
-					</time>
-				)
-			}
+					{
+						post.frontmatter.date && (
+							<time className="post-meta">
+								{formatDate(post.frontmatter.date)}
+							</time>
+						)
+					}
 				</header>
 
 				<article className="post-content"
 					dangerouslySetInnerHTML={{__html: post.html}}></article>
 			</section>
-		);
-	}
+		</Layout>
+	);
 
-	componentDidMount() {
-		this.createCarousels();
-	}
+}
 
-	componentDidUpdate() {
-		this.createCarousels();
-	}
-
-	createCarousels() {
-		let carousels = Array.from(document.querySelectorAll('.carousel'));
-		for (let carousel of carousels) {
-			new Flickity(carousel, {
-				cover: true,
-				imagesLoaded: true
-			});
-		}
+function createCarousels() {
+	let carousels = Array.from(document.querySelectorAll('.carousel'));
+	for (let carousel of carousels) {
+		new Flickity(carousel, {
+			cover: true,
+			imagesLoaded: true
+		});
 	}
 }
 
