@@ -1,45 +1,40 @@
 /* global graphql */
-import React, { Component } from 'react';
-import PropTypes from 'prop-types';
-import Link from 'gatsby-link';
+import React from 'react';
+import { Link } from 'gatsby';
+
+import Layout from '../components/Layout.js';
 
 import { formatDate } from '../utils.js';
 
-export default class Blog extends Component {
-	static propTypes = {
-		data: PropTypes.object.isRequired
-	};
+export default function Blog({ data: { allMarkdownRemark: { edges } } }) {
+	const posts = edges.filter(edge =>
+		edge.node.frontmatter.type === 'post'
+	);
 
-	render() {
-		const posts = this.props.data.allMarkdownRemark.edges.filter(edge =>
-			edge.node.frontmatter.type === 'post'
-		);
+	const postListItems = posts.map(post => (
+		<li key={post.node.fields.slug}>
+			<span className="post-meta">
+				{formatDate(post.node.frontmatter.date)}
+			</span>
+			<h2>
+				<Link to={post.node.fields.slug}
+						className="post-link">
+					{post.node.frontmatter.title}
+				</Link>
+			</h2>
+		</li>
+	));
 
-		const postListItems = posts.map(post => (
-			<li key={post.node.fields.slug}>
-				<span className="post-meta">
-					{formatDate(post.node.frontmatter.date)}
-				</span>
-				<h2>
-					<Link to={post.node.fields.slug}
-							className="post-link">
-						{post.node.frontmatter.title}
-					</Link>
-				</h2>
-			</li>
-		));
-
-		return (
-			<div>
-				<h1 className="page-heading">
-					Blog
-				</h1>
-				<ul className="post-list">
-					{postListItems}
-				</ul>
-			</div>
-		);
-	}
+	return (
+		<Layout>
+			<h1 className="page-heading">
+				Blog
+			</h1>
+			<ul className="post-list">
+				{postListItems}
+			</ul>
+		</Layout>
+	);
 }
 
 export const pageQuery = graphql`
